✨ feat: connecting auth to backend, wip: connecting main data to backend (I hate states)

+++ cmd/root.go
@@ -38,7 +38,7 @@ func init() {
 	rootCmd.PersistentFlags().StringVar(&cfg.DBConnectionString, "db-connection-string", "./shortener.db", "Connection string used to connect to database")
 	rootCmd.PersistentFlags().StringVar(&cfg.DBEngine, "db-engine", "sqlite3", "Engine for database should be one of (postgres, mysql, sqlite3, mssql)")
 	rootCmd.PersistentFlags().StringVar(&cfg.JWTSecret, "jwt-secret", "superdupersecret", "Secret used to sign JWT tokens")
-	rootCmd.PersistentFlags().IntVar(&cfg.TokenExpireTime, "token-expire-time", 5, "How much a token could be valid (In hour)")
+	rootCmd.PersistentFlags().Int64Var(&cfg.TokenExpireTime, "token-expire-time", 5, "How much a token could be valid (In hour)")
 	rootCmd.PersistentFlags().BoolVar(&cfg.Debug, "debug", false, "Weather to enable debug log or not (will print sensitive data)")
 }
